requirements.txt
picamera
Pillow